+++ Plugins/Experimental/Experimental.cpp
@@ -1,6 +1,7 @@
 #include "Experimental.hpp"
 #include "Experimentals/SuppressPlayerLoginInfo.hpp"
 #include "Experimentals/AdjustReputationFix.hpp"
+#include "Experimentals/DisableLevelUpValidation.hpp"
 
 #include "Services/Config/Config.hpp"
 
@@ -30,6 +31,12 @@ Experimental::Experimental(Services::ProxyServiceList* services)
     {
         LOG_INFO("EXPERIMENTAL: Attempting to resolve faction/reputation crash.");
         m_AdjustReputationFix = std::make_unique<AdjustReputationFix>(GetServices()->m_hooks.get());
+    }
+
+    if (GetServices()->m_config->Get<bool>("DISABLE_LEVELUP_VALIDATION", false))
+    {
+        LOG_INFO("EXPERIMENTAL: Disabling LevelUp Validation.");
+        m_DisableLevelUpValidation = std::make_unique<DisableLevelUpValidation>(GetServices()->m_hooks.get());
     }
 }